carbonsearch: string -> int{32, 64} for tag/metric/join key
This commit introduces types for tags, metrics, and joins. These types
are numeric, and generated by hashing the strings. split indexes no longer
maintain any of the original data: they only have the hashed
representations.

As a result, the query path maps only contain (and hash) int32/64s,
rather than strings: so memory usage is reduced and the hashing functions have
fewer bytes to hash. It should also simplify keeping the indexes as
sorted slices when we move that way.

The hash -> metric mapping is kept in the Database, and
only applied as a last-minute transform once the query is almost
complete.

This results in (much) better benchmarks in split_test.go, but seemingly
no improvement in wrk2 results.