carbonsearch: mitigate birthdays by expanding join keys to uint64
see "birthday paradox": we have ~30k join keys, so even with a space of
uint32 there's a 1% chance of collision around 9300 join keys. Bumping
to uint64 raises the threshold for 1% collision probability to around
300m members, which is comfortably above either our likely join key
population or our total metric population.

thanks to Nate and Damian for feedback/discussion on this point