carbonsearch: basic trigram index for text search on metric names
This is WIP, since the text index does not respect the relative
positioning of trigrams in the user's query. Thus, a query like: "cron$"
will match a document like "ron.crocodile.option" because "cro", and
"ron", and "on$" are all present in the document. There's already
a (commented out) test for this case, so that'll be the next revision.

Borrows liberally from https://github.com/dgryski/go-trigram.