carbonsearch: graphite-shaped search queries
this takes a 'metric' like:

    virt.v1.tag-pool:www.lb-live:true.server-state:installed

and returns metrics which were generated by hosts which match these
characteristics right now (using AND semantics).

This is an in-memory service that will run beside the carbonzipper,
consuming from a handful of Kafka topics to populate the index.

There's still some work to do, in particular:

1) monitoring/syslogging
2) simple snapshotting persistence (at least for the 'full' index, often
added by humans), and
3) some anti-entropy converging pressure of some kind

...but it's complete enough to build indexes and serve search queries
from them.

This allows us to query metrics by attributes which are stored external
to the graphite namespace hierarchy.