tags: omit needless detail
As far as I can see right now, we should never need the raw value for the
'value' portion of a tag. We need a unique identifier for the 'key' so that we
can know that we need to replace the hash which represents the previous
iteration of this key. In doing so, we don't need the raw strings: a hash of
'service-key' ought to do.

This also shuffles around the tag parsing logic slightly.