database & kafka consumer: defend against malformed messages
There's something putting log-formatted messages (timestamp, etc) into the kafka topics,
and the symptom is that Msg decodes into a struct full of empty strings.

The correct place for this logic to live is purely in the database, but
allowing the malformed messages to pass through the kafka consumer and
then logging when database returns an error will generate huge amounts
of noise until we stop sending the malformed messages, so
double-validate for now.