split index: ensure join lists are sorted
Previously they were materialized in map iteration order. This was (generally)
successful in tests, since the set populations were small enough (sometimes as
small as 1) that this was hard to bump into.

Unsorted sets of any kind break the various union/intersect algorithms, and as
such give very weird results when used with those algorithms.