database test: tests to catch missing sorts on materialize/ingest
This is a tricky topic, since these tests are probabilistic: they fail based on
map iteration order. This happens because the test sets are fairly small, and
so they can be materialized into the read index in "sorted enough" order to
allow tests to pass.

In order to compensate for this, regenerate/retest 25 times: this reshuffles
the ordering many times, and usually yields ~4-8 failures per run.

Also tweak the input order of some test metrics to expose the case where
metrics are not sorted on ingest.