bloom index/database: move metric string mapping to a now-generational bloom index
This commit removes responsibility for maintaining the index.Metric -> string
map from Database and adds it to the Bloom index. This allows this mapping to
be part of the materialization, and so the mapping will remain in sync with the
metrics that are queryable by text.

This also means that the last remaining lock in the query path is a trivial one
for fetching the correct split index.

As part of writing bloom.Materialize, some metric reporting was also added
(comparable to what's in the split indexes).