main/database: statically create/assign split indexes from config
This removes a source of unbounded growth in the face of bad user input:
blithely accepting messages and creating indexes for them dynamically.

It seems reasonable that the total number of indexes will be small (<10), so
writing out the assignments in a config file shouldn't be too much of a burden.

Leaning on a config file should also make the system easier to reason about.

As a side bonus, it removes some locks from database and the query path.