split/full index: empty/missing query tags -> empty set
This enforces the behavior that a tag like "servers-blorg:bazbazbazbaz",
which is entirely bogus and not being indexex by any data source, will
yield 0 results when included in a query.

Previously entirely unknown tags (never seen) were simply ignored by the
split index, which meant that a query like
'servers-dc:us_west.servers-os:co3' would return results for all of the
'us_west' DC, even though none of them (hopefully :)) were running
Centos 3.

The full index was actually already doing this, but with some slightly
clever implicit behavior: because the metricSets [][]index.Metric was
intialized with a length equal to the submitted query tags, any tag with
0 results would result in the zero value for that position in the slice,
which would be an empty slice. When intersected, this would produce an
empty set. This was standardized to work the same way (early return) as
the split index.