database: add a table of contents for humans
This is a deeply inefficient tool for helping people understand what indexes,
services, keys, and values are out there.

This isn't yet synced with index generations, so things won't be 100% in
line with what's really available to query (usually it'll be running
ahead somewhat -- it accepts writes as they come in, while the indexes
are on a materialization schedule).