database: add Autocomplete
This function will take a user's partial tag and attempt to provide options for
finishing it based on values found in the table of contents.

This is a pretty simple-minded implementation that just iterates over the whole
search space doing strings.HasPrefix.

If this becomes a performance issue (it read-locks the table of contents, which
will block the write path for incoming tags/metrics), a radix trie of tags
would probably be fairly effective, particularly if 'service', 'key', and
'value' portions are distinct in the trie.

Alternatively (or in combination) we could switch the table of contents to
materialize on the same schedule as the indexes, which would have other
benefits (it would always be in sync with the indexes, whereas now it is
persistently a little bit ahead).