table of contents/autocomplete: track full index
The autocompleter and table of contents will now include metrics/tags from the
full index. In order to avoid a pile of code duplication, a fair bit of
factoring was done on the table of contents code: this introduced an indexEntry
interface. This code is all fairly sloppy and "just muddle through", so any
suggestions on how to structure it more nicely would be most welcome.