bloom: add RW mutexes around the swappable indexes
The atomic usage was unsound: it works in the split index because we discard
the old value on each materialize, but in the bloom index we 'catch up' an
older bloom in order to reduce materialization time. As a result, concurrent
read/write was possible on the docToMetric map.

This is a bit of a stopgap, as we're planning to shift towards
https://github.com/dgryski/go-postings at some point for the space savings (and
consequent ability to keep more previous generations around in memory).