database/tag: change query parser to enable quoted tag values
This allows longer chunks of metric names to be used in 'text-match' queries.

For now the quote characters are '<' and '>', but this is straightforward to
change.

An alternate implementation would be simply escaping '.' with \ or similar.
However, graphite uses this same syntax, and performs the escaping in its
parser. As a result, it isn't clear that we can rely on the escape \ being
passed through without something goofy like stacks of \\\\\\\. Quoting is
probably a reasonable approach since it allows some freedom for extensions of
the matching syntax.