main/consumers: add 'WaitUntilWarm'
Each consumer can now block the index from serving queries until it is
considered 'warm'. What 'warm' means varies per consumer: for Kafka it
means that for all 3 topics, the average progress of the partitions
within a topic (measured as currentOffset/highWaterMark) is >=
warm_threshold. For the HTTP consumer, an endpoint is exposed
(/consumer/progress) -- a number can be sent to that endpoint in order
to change the HTTP consumer's idea of how warm it is.

The index will serve search queries once all configured consumers
consider themselves warm.