main/consumers: Fix HTTP consumer's interaction with graceful restart
This removes a problem in earlier iterations where the HTTP consumer would not
yield the port to the new process. This patch introduces the usage of
'ServeWithOptions' and gracehttp.StartupHook from a patched
facebookgo/gracehttp. This API might change in discussion with upstream, but it
allows us to be sure to shut off the previous HTTP consumer before trying to
start the new one.

We couldn't use a simpler option (like having the new process just poll for
port availability) because the second process isn't ready to serve traffic
until it has warmed up the index, which requires the HTTP consumer API to be
working.