kafka consumer: improve progress reporting
Previously it was using the raw offsets to calculate progress, but these
numbers were gigantic: 999990000 / 100000000 is a rather nice
percentage, but it doesn't represent progress of this carbonsearch
instance since it started. Instead, calculate progress relative to the
first offset seen by this consumer. Additionally improve the warmup
progress reporting slightly.

Also a few small tweaks to handle topics with zero messages.