main: reshuffle things so that the HTTP handlers can be tested
Mostly this involved using closures instead of global variables. I realized
later that I could just assign to the globals in the tests (as, indeed, the
timeBucket things still do, essentially), but I like the idea of being able to
run tests in parallel on different ports in principle.

There are some sketches of tests for the findHandler at present: I'd like to
refactor this to be a bit less code duplication, and to also exercise the HTTP
consumer.