Revert "document: switch to trigrams"
This reverts commit 5e8c2ef22e0269ce6f1751a032c7b1785110e4b5.

The experiment with a trigram-specific index in go-postings (by preallocating
a giant array for each slot) led to substantial pressure on the Go GC. While
the array-based index was (as expected) much faster for indexing, OOM is
a significant downside.

Since that experiment failed, revert to the better selectivity of quadgrams.