database, index/text: Fix tests to cover text index configuration bug
Default name for the text service is "text", i.e. as in
"text-match:<metric.name>". The service name is set in a configuration
file via the `text_index_service` key.

When setting `text_index_service` to custom value, carbonsearch still
under some circumstances expects "text". Apparently, "text" is hardcoded
somewhere in the source code. That's a bug.

    $ grep text_index_service carbonsearch.yaml
    text_index_service: "loltxt"

    $ curl -s -X GET 'localhost:8070/metrics/find/?format=json;query=virt.v1.*.*' | jq '.matches[].path'
    "virt.v1.*.loltxt-"

    $ curl -s -X GET 'localhost:8070/metrics/find/?format=json;query=virt.v1.*.loltxt-match:wat'
    database: error while querying index text index: text index Query: no text searches in query: [loltxt-match:wat]

    $ curl -s -X GET 'localhost:8070/metrics/find/?format=json;query=virt.v1.*.text-match:wat'
    {"name":"virt.v1.*.text-match:wat"}

This patch changes the test suite to cover the case wheh text service
name is set to a non-default value ("tekst" in this case).