database, index/text: Respect text_index_service config value everywhere
As described in the previous commit, carbonsearch allows to configure a
name for the text service: "text" in "text-match:<some.thing>" is
defined by the `text_index_service` configuration variable. But a
default "text" string was hardcoded in some code instead of using the
config value. This patch fixes the bug and makes the test suite pass
again.